Rev3 Chassis Initial Structure and CAD
Please download and save all these files locally, and confirm with anyone else if they are working on the same assembly. Once you are done, save as a SLDASM and select pack and go. You can either use a zip file or just upload all the files that you packed and went with. Also, please do remember to add the revision number to the file name as well as the commit title. If possible, please add the date as well. Describe what was updated in the assembly in the commit description.